+++ src/main/java/de/uni_mannheim/informatik/dws/wdi/ExerciseIdentityResolution/Comparators/author/BookAuthorComparatorPreprocessedJaro.java
@@ -52,7 +52,7 @@ public class BookAuthorComparatorPreprocessedJaro implements Comparator<Book, At
         }
 
         // postprocessing
-        int postSimilarity = 1;
+        double postSimilarity = 1;
         if (maxSimilarity <= 0.3) {
             postSimilarity = 0;
         }
